Stripped excess out of ontology to move to subclasses.
